requirements.txt
bcrypt==5.0.0
blinker==1.9.0
cachelib==0.13.0
click==8.3.1
colorama==0.4.6
Flask==3.1.2
Flask-Bcrypt==1.0.1
Flask-Session==0.8.0
Flask-SQLAlchemy==3.1.1
greenlet==3.3.1
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.3
msgspec==0.20.0
SQLAlchemy==2.0.46
typing_extensions==4.15.0
Werkzeug==3.1.5
